+++ apis/externalsecrets/v1/secretstore_yandex_types.go
@@ -27,3 +27,19 @@ type YandexAuth struct {
 type YandexCAProvider struct {
 	Certificate esmeta.SecretKeySelector `json:"certSecretRef,omitempty"`
 }
+
+// ByID configures the provider to interpret the `data.secretKey.remoteRef.key` field in ExternalSecret as secret ID.
+type ByID struct{}
+
+// ByName configures the provider to interpret the `data.secretKey.remoteRef.key` field in ExternalSecret as secret name.
+type ByName struct {
+	// The folder to fetch secrets from
+	FolderID string `json:"folderID"`
+}
+
+// +kubebuilder:validation:MinProperties=1
+// +kubebuilder:validation:MaxProperties=1
+type FetchingPolicy struct {
+	ByID   *ByID   `json:"byID,omitempty"`
+	ByName *ByName `json:"byName,omitempty"`
+}
